refactor(component):폴더명 변경
